backend/requirements/base.txt
# Core Django
Django==5.0.6
djangorestframework==3.15.2
django-environ==0.11.2

# Django Channels (Real-Time / WebSockets)
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.2

# Celery (Async Tasks)
celery==5.4.0
django-celery-beat==2.6.0
django-celery-results==2.5.1
redis==5.0.6
flower==2.0.1

# Database
psycopg[binary]==3.1.19
django-redis==5.4.0

# GeoDjango / PostGIS
GDAL==3.8.4

# Authentication
djangorestframework-simplejwt==5.3.1
social-auth-app-django==5.4.1

# API
django-filter==24.2
drf-spectacular==0.27.2
django-cors-headers==4.4.0

# File Storage (AWS S3)
django-storages[s3]==1.14.3
boto3==1.34.110

# Phone / OTP
twilio==9.1.1

# Payments
stripe==9.9.0

# Rate Limiting
django-ratelimit==4.1.0

# Security & Encryption
django-encrypted-model-fields==0.6.5
cryptography==42.0.7

# Admin
django-unfold==0.21.0
django-two-factor-auth==1.16.0

# Search
django-elasticsearch-dsl==8.0

# Logging
structlog==24.2.0

# Code Quality (dev tools in dev.txt)
